% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.r
\name{download_maralago}
\alias{download_maralago}
\title{Download the full Excel file}
\usage{
download_maralago(destfile, overwrite = TRUE)
}
\arguments{
\item{destfile}{A character string with the name where the downloaded file is saved.
Tilde-expansion is performed.}

\item{overwrite}{Will only overwrite existing path if TRUE.}
}
\description{
Download the full Excel file
}
