% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maRalago-package.R
\docType{package}
\name{maRalago}
\alias{maRalago}
\alias{maRalago-package}
\title{Is Trump At Mar-a-Lago (and more)}
\description{
See if Trump is at Mar-a-Lago and also retreive total trips/costs to Mar-a-Lago
as well as a list of programs with unit costs to assist in derivation of what the
narcisstic waste of money could to to fund or keep funding.
}
\details{
It relies on a Google spreadsheet maintaind by the Center for American Progress Action
Fund which was discovered after seeing \url{https://istrumpatmaralago.org/}.
}
\author{
Bob Rudis (bob@rud.is)
}
